requirements.txt
aws-cdk.core==1.128.0
aws-cdk.aws-ec2==1.128.0
aws-cdk.aws-eks==1.128.0
aws-cdk.aws-apigateway==1.128.0
aws-cdk.aws-dynamodb==1.128.0
aws-cdk.aws-s3==1.128.0
aws-cdk.aws-iam==1.128.0
boto3==1.18.0